Extract CheckTarget into GameLogic::Combat
Move CheckTarget (resolves the attack-target tile from selection/cursor) out of
ZzzInterface.cpp into GameLogic/Combat/CombatTarget.{h,cpp} under namespace
GameLogic::Combat. Its callers are the combat skill files (in-namespace, just
need the include) and SummonSystem (now qualified). No behaviour change.
