+++ src/source/GameLogic/Combat/CombatTarget.cpp
@@ -1,0 +1,76 @@
+#include "stdafx.h"
+#include "GameLogic/Combat/CombatTarget.h"
+
+// Includes mirror ZzzInterface.cpp, the unit this was extracted from.
+#include <imm.h>
+#include "UI/Legacy/UIManager.h"
+#include "Render/Textures/ZzzOpenglUtil.h"
+#include "Render/Models/ZzzBMD.h"
+#include "Render/Terrain/ZzzLodTerrain.h"
+#include "Engine/Object/ZzzInfomation.h"
+#include "Engine/Object/ZzzObject.h"
+#include "Engine/Object/ZzzCharacter.h"
+#include "Engine/Object/PlayerActionState.h"
+#include "Render/Textures/ZzzTexture.h"
+#include "Engine/AI/ZzzAI.h"
+#include "Engine/Object/ZzzInterface.h"
+#include "UI/Chat/ChatInput.h"
+#include "GameLogic/Combat/ClassAttack.h"
+#include "GameLogic/Combat/SkillCast.h"
+#include "GameLogic/Combat/SkillExecution.h"
+#include "Core/Input/ImeInput.h"
+#include "Engine/Object/ZzzInventory.h"
+#include "Engine/Object/ZzzOpenData.h"
+#include "Render/Effects/ZzzEffect.h"
+#include "Scenes/SceneCore.h"
+#include "Engine/Pathing/ZzzPath.h"
+#include "Audio/DSPlaySound.h"
+#include "I18N/All.h"
+#include "GameLogic/Events/MatchEvent.h"
+#include "GameLogic/Items/PersonalShopTitleImp.h"
+#include "GameLogic/Quests/CSQuest.h"
+#include "GameLogic/Items/CSItemOption.h"
+#include "GameLogic/NPCs/npcBreeder.h"
+#include "GameLogic/Pets/GIPetManager.h"
+#include "Character/CSParts.h"
+#include "UI/Legacy/UIMapName.h"	// rozy
+#include "GameLogic/Events/Cinematic/CDirection.h"
+#include "World/MapInfra/MapManager.h"
+#include "GameLogic/Events/Event.h"
+#include "UI/NewUI/NewUISystem.h"
+#include "GameLogic/Events/w_CursedTemple.h"
+#include "UI/Legacy/UIControls.h"
+#include "GameLogic/Social/PartyManager.h"
+#include "UI/NewUI/Dialogs/NewUICommonMessageBox.h"
+#include "GameLogic/Skills/SummonSystem.h"
+#include "GameLogic/Skills/SkillManager.h"
+#include "World/MapInfra/w_MapHeaders.h"
+
+extern int SelectedCharacter, TargetX, TargetY;
+extern bool SelectFlag;
+
+namespace GameLogic::Combat
+{
+bool CheckTarget(CHARACTER* c)
+{
+    if (SelectedCharacter >= 0 && SelectedCharacter < MAX_CHARACTERS_CLIENT)
+    {
+        TargetX = (int)(CharactersClient[SelectedCharacter].Object.Position[0] / TERRAIN_SCALE);
+        TargetY = (int)(CharactersClient[SelectedCharacter].Object.Position[1] / TERRAIN_SCALE);
+        VectorCopy(CharactersClient[SelectedCharacter].Object.Position, c->TargetPosition);
+        return true;
+    }
+    else
+    {
+        RenderTerrain(true);
+        if (SelectFlag)
+        {
+            VectorCopy(CollisionPosition, c->TargetPosition);
+            TargetX = (int)(c->TargetPosition[0] / TERRAIN_SCALE);
+            TargetY = (int)(c->TargetPosition[1] / TERRAIN_SCALE);
+            return true;
+        }
+    }
+    return false;
+}
+}
